Update dockstar kernel to 3.14.2
Since this kernel version only supports FDT bindings for the Dockstar
(the "real" [setup] code has been removed), switch to FDT. And since
I don't want to update the bootloader on the device, I use the feature
of the kernel to append the compiled FDT after the zImage.

For now, I switched ptxdist to zImage and generate the uImage with the
appended device tree manually with a script.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>